Escaped semicolons, changed generated file dir
The semicolons in the content string in write_gtest_main were not
appearing in the written file; escaping them with backslashes fixed the
issue

I used to create a folder, generated, inside of the source directory for
main.cpp. I now create it inside of CMAKE_BINARY_DIR isntead, and gave
it the slightly more descriptive naem generated_code.