Added project command, fixed include bug
I added a project command to the example directory because it's a good
practice and it is a convenient place to stick a version number.

When I changed the name of the root CMakeLists.txt to add-gtest.cmake, I
failed to update the include directory in the example CMakeLists.txt.
This has been fixed.